pythonnetapp/requirements.txt
Flask
flask_sqlalchemy
requests
six
Flask-RESTful
psycopg2-binary --no-binary :all: